investors avg profit loss to double type

+++ lib/features/top_investors/data/model/investor_holding_model.dart
@@ -30,7 +30,7 @@ class InvestorMetricsModel extends InvestorMetrics {
       totalTrades: json['totalTrades'],
       profitableTrades: json['profitableTrades'],
       lossTrades: json['lossTrades'],
-      avgProfitLossRatio: json['avgProfitLossRatio'],
+      avgProfitLossRatio: json['avgProfitLossRatio'].toDouble(),
     );
   }
